requirements.txt

opencv-python-headless
pillow
tk
